+++ system-tests/tia-maven/maven-project/sub-project-A/src/test/java/org/example/NestedTest.java
@@ -1,0 +1,23 @@
+package org.example;
+
+import org.junit.jupiter.api.Nested;
+import org.junit.jupiter.api.Test;
+
+public class NestedTest {
+
+	@Nested
+	class Child1 {
+		@Test
+		void getDefaultParameter() {
+			new SUTA().bla();
+		}
+	}
+
+	@Nested
+	class Child2 {
+		@Test
+		void getDefaultParameter() {
+			new SUTA().foo();
+		}
+	}
+}